Clear things up with `git reset`
It doesn't make sense to mention `git reset --hard`
during the staging explanation, where we don't know a
thing about commits yet.

On the other hand, collaboration/advanced is too late
since we use that command before. During 'commits' it is
too early since we need to learn about 'branches' and
'smart pointers' with HEAD before.

Therefore it is now moved into the branching chapter in the end.