Enhance workflows; add GitLab issue/milestone for merge request
This splits the Forking workflow for GitHub and GitLab.
While it is common for GitHub and social coding to always
fork the repositories, own GitLab installations with roles and
permissions may just use local branches for merge requests.

This also is reflected by updated images.

In addition to this, users will now create a milestone
and issue which is automatically resolved later in the merge
request lab.

refs #10
fixes #13